Add basic untested logic for config whitelist
The idea here is to retrieve as many configuration options as possible
without ever retrieving anything that could be including some sensitive
information with PII or application secrets.

You could say that technically Rails config should be secret-free since
secrets.yml was introduced but that may not be the case.

In the long-run it might make sense to offer a way to scan the app for
high entropy strings and potentially private stuff like URIs.